feat: tweak packages, move some files around that needed movin
