+++ app/src/main/java/com/mnbpdx/gardenbook/ui/screens/detail/DetailScreenViewModel.kt
@@ -1,4 +1,4 @@
-package com.mnbpdx.gardenbook
+package com.mnbpdx.gardenbook.ui.screens.detail
 
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
