+++ app/src/main/java/com/mnbpdx/gardenbook/ui/screens/home/HomeScreenViewModel.kt
@@ -1,4 +1,4 @@
-package com.mnbpdx.gardenbook.ui.home
+package com.mnbpdx.gardenbook.ui.screens.home
 
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
