+++ app/src/main/java/com/mnbpdx/gardenbook/ui/screens/loading/LoadingScreen.kt
@@ -1,4 +1,4 @@
-package com.mnbpdx.gardenbook.ui.loading
+package com.mnbpdx.gardenbook.ui.screens.loading
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Box
